requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
gspread==6.0.2
oauth2client==4.1.3